Placed function in correct controller, changed Arena and Booking Models (minor changes to the router)
